feat(present): 质量灯 (五轴两轴进浏览器) + Rule 24 依从性提升 (funding 73→87.5%)
质量灯 (方向 B):
- eval-core.js: 离线 scorer 的纯数字核抽到 src (浏览器可跑, scripts 侧 re-export,
  行为逐字节一致 — harness 73/73 + qbr 94.4% 复核不变)
- quality-lights.js: assessSlot/assessDeck — 视觉审计 (TEXT_OVERFLOW/OVERLAP/…)
  + 数字接地 (字面 + Rule 24 衍生引用), 零 LLM 纯本地
- author-2d 每页 🟢🟡🔴 徽章 (renderDeck 尾挂, 非阻塞), hover 见明细, 点击进状态栏;
  接地源 = prompt 文本 + 整本 outline (重开的 deck 也有 — shared.slides 随身)
- e2e 真阳性: demo fixture 的 420/310 确实不在 prompt 里, 灯诚实地红了

Rule 24 依从性:
- worked example B 加一行引用示范 (trend="-25% (2.8 vs 2.1)") — 最小 token 成本
- 重烤 funding-round: precision 73→87.5%, entity recall 67→83; 残留一个裸 43%
  由常备语料继续追踪

测试: test-quality-lights 6 断言进 runner; npm test 136/136; corpus mean 97.4
