+++ scripts/run-tests.mjs
@@ -215,6 +215,7 @@ const TESTS = [
   { category: 'present', file: 'sdf-js/scripts/test-decor.mjs' },
   { category: 'present', file: 'sdf-js/scripts/test-deck-io.mjs' },
   { category: 'present', file: 'sdf-js/scripts/test-deck-contract.mjs' },
+  { category: 'present', file: 'sdf-js/scripts/test-quality-lights.mjs' },
 ];
 
 // -----------------------------------------------------------------------------
