+++ sdf-js/apps/present/author-2d.html
@@ -246,6 +246,19 @@
       .slide-card.busy .slide-busy {
         display: flex;
       }
+      /* Sprint 67: per-slide quality light (visual audit + number grounding) */
+      .qlight {
+        position: absolute;
+        top: 8px;
+        left: 8px;
+        z-index: 5;
+        border: none;
+        border-radius: 999px;
+        padding: 3px 7px;
+        background: rgba(20, 22, 28, 0.82);
+        font-size: 11px;
+        cursor: pointer;
+      }
       #keyrow {
         position: fixed;
         right: 14px;
