+++ sdf-js/examples/scaffold-pipeline/eval-funding-round-slidedata.json
@@ -3,116 +3,116 @@
     "title": "Redo Announces $81 Million Series B to Support Commerce Technology Expansion",
     "body": [
       "Draper, Utah, June 23, 2026",
-      "Series B financing announcement",
-      "Commerce technology platform"
+      "Commerce technology platform secures major funding round",
+      "Led by Smash Capital with existing investor participation"
     ]
   },
   {
     "title": "Agenda",
     "body": [
-      "Financing details and valuation",
-      "Current platform adoption",
-      "Strategic funding priorities",
-      "AI-driven product development",
-      "Network infrastructure capabilities",
-      "Acquisition and international expansion"
+      "Financing details and valuation milestone",
+      "Current platform adoption and merchant base",
+      "Leadership perspectives on market opportunity",
+      "Strategic use of capital for AI development",
+      "Network infrastructure and operations model",
+      "ReturnBear acquisition and international expansion"
     ]
   },
   {
-    "title": "Series B Financing Details",
+    "title": "Series B Financing Overview",
     "body": [
-      "$81 million Series B financing",
-      "$1.25 billion valuation",
+      "Redo raised $81 million in Series B funding",
+      "The financing values the company at $1.25 billion",
       "Round led by Smash Capital",
-      "Participation from existing investors Pelion Venture Partners and Cervin Ventures"
+      "Existing investors Pelion Venture Partners and Cervin Ventures participated"
     ]
   },
   {
-    "title": "Platform Adoption Metrics",
+    "title": "Platform Adoption Scale",
     "body": [
       "More than 4,100 brands use the Redo platform today",
       "Over 1,750 merchants use multiple Redo offerings across the suite",
-      "Demonstrates cross-platform engagement and product suite adoption"
+      "Platform demonstrates cross-product engagement with merchant base"
     ]
   },
   {
-    "title": "Mission and Strategic Vision",
+    "title": "CEO Vision: Merchant-Shopper Relationships",
     "body": [
-      "Every merchant is fighting to own the relationship with their shoppers, said Sterling Snow, co-founder and CEO",
-      "This funding allows us to accelerate that mission",
-      "Build the infrastructure brands need to own shopper relationships"
+      "Sterling Snow serves as co-founder and CEO of Redo",
+      "Snow states every merchant is fighting to own the relationship with their shoppers",
+      "Funding allows acceleration of that mission and building infrastructure brands need"
     ]
   },
   {
-    "title": "Investor Perspective on Market Opportunity",
+    "title": "Investor Thesis on Market Expansion",
     "body": [
-      "Redo started by building trust with merchants through post purchase experiences, said Paul Szurek of Smash Capital",
-      "Has expanded into a much larger opportunity",
-      "Validation of platform evolution beyond initial product"
+      "Paul Szurek represents lead investor Smash Capital",
+      "Redo started by building trust with merchants through post purchase experiences",
+      "Company has expanded into a much larger opportunity beyond initial focus"
     ]
   },
   {
     "title": "AI-Driven Commerce Tools Development",
     "body": [
       "New capital will fund development of AI-driven commerce tools",
-      "Product roadmap focused on personalization and automation",
-      "Three distinct AI product initiatives planned"
+      "Investment focuses on next-generation shopping and marketing capabilities",
+      "Suite designed to enhance merchant-customer engagement"
     ]
   },
   {
     "title": "Agentic Shopping Experiences",
     "body": [
-      "Agentic shopping experiences tailored to individual shoppers",
-      "AI agents designed to personalize the shopping journey",
-      "Focus on individual customer customization"
+      "Platform will include agentic shopping experiences",
+      "Experiences tailored to individual shoppers",
+      "AI agents personalize the shopping journey"
     ]
   },
   {
-    "title": "Marketing Agent Platform",
+    "title": "Marketing Agent for Campaigns",
     "body": [
-      "Marketing agent for personalized email and text campaigns",
-      "Automated personalization across communication channels",
-      "Designed to enhance merchant-customer engagement"
+      "Marketing agent capability for personalized email campaigns",
+      "Agent also handles personalized text campaigns",
+      "Automation targets customer communication at scale"
     ]
   },
   {
     "title": "Post-Purchase Concierge Service",
     "body": [
-      "Post-purchase concierge service designed to increase customer lifetime value",
-      "Focus on retention and value maximization after initial purchase",
-      "Extension of post-purchase experience foundation"
+      "Post-purchase concierge service under development",
+      "Service designed to increase customer lifetime value",
+      "Focuses on relationship extension beyond transaction"
     ]
   },
   {
     "title": "Returns Network Infrastructure",
     "body": [
-      "Our network allows customers to return products locally, said Aaron Evett, chief commerce officer",
-      "We can inspect, refurbish, re-fulfill, and reship inventory",
-      "Local returns processing creates operational efficiency and inventory recovery"
+      "Aaron Evett serves as chief commerce officer",
+      "Network allows customers to return products locally",
+      "Company can inspect, refurbish, re-fulfill, and reship inventory through network"
     ]
   },
   {
     "title": "ReturnBear Acquisition",
     "body": [
-      "Redo announced the acquisition of ReturnBear",
-      "Extends returns servicing coverage to more than 100 countries",
-      "Significantly expands international returns infrastructure"
+      "Redo announced the acquisition of ReturnBear alongside the financing",
+      "Acquisition extends returns servicing coverage to more than 100 countries",
+      "Deal expands geographic reach of returns infrastructure"
     ]
   },
   {
     "title": "International Expansion Plans",
     "body": [
-      "Plans for international expansion of merchant network",
-      "Global scaling strategy alongside ReturnBear acquisition",
-      "Extends platform availability beyond domestic market"
+      "Company outlined plans for international expansion of its merchant network",
+      "Expansion strategy follows ReturnBear acquisition",
+      "Growth aims to serve merchants and customers across global markets"
     ]
   },
   {
     "title": "Strategic Outlook",
     "body": [
-      "Funding supports transition from post-purchase specialist to comprehensive commerce platform",
-      "International expansion and AI development as dual growth drivers",
-      "Network effects from returns infrastructure and multi-product adoption"
+      "Redo positioned to accelerate commerce infrastructure development",
+      "Combined funding, acquisition, and AI investments support merchant ownership of customer relationships",
+      "Platform expansion targets comprehensive post-purchase and engagement solutions at global scale"
     ]
   }
 ]
