+++ sdf-js/examples/scaffold-pipeline/eval-funding-round/slots/slot-04-challenges.json
@@ -3,8 +3,8 @@
   "slotName": "challenges",
   "slotTitle": "Challenges",
   "slotPurpose": "What did not go well and what we learned",
-  "sourceSlideIdx": 10,
-  "sourceSlideTitle": "Returns Network Infrastructure",
+  "sourceSlideIdx": 11,
+  "sourceSlideTitle": "ReturnBear Acquisition",
   "sceneData": {
     "name": "investor-update/challenges",
     "layout": "row",
@@ -21,45 +21,28 @@
         }
       },
       {
-        "type": "fishbone",
+        "type": "bullet-list",
         "x": 40,
         "y": 160,
         "w": 1200,
         "h": 520,
         "args": {
-          "effect": "Returns Network Scaling",
-          "branches": [
+          "title": "ReturnBear Acquisition Context",
+          "items": [
             {
-              "label": "Infrastructure",
-              "causes": [
-                "Local processing nodes",
-                "Inspection facilities",
-                "Refurbishment capacity"
-              ]
+              "icon": "handshake",
+              "label": "Acquired ReturnBear with financing round",
+              "status": "done"
             },
             {
-              "label": "Operations",
-              "causes": [
-                "Inventory recovery time",
-                "Re-fulfillment logistics",
-                "Quality control standards"
-              ]
+              "icon": "globe-hemisphere-west",
+              "label": "Returns coverage extended to 100+ countries",
+              "status": "done"
             },
             {
-              "label": "Network",
-              "causes": [
-                "Geographic coverage gaps",
-                "Partner integration",
-                "Customer access points"
-              ]
-            },
-            {
-              "label": "Efficiency",
-              "causes": [
-                "Process optimization",
-                "Turnaround speed",
-                "Cost per unit"
-              ]
+              "icon": "arrows-out",
+              "label": "Geographic reach of infrastructure expanded",
+              "status": "done"
             }
           ]
         }
@@ -76,20 +59,20 @@
     ],
     "forbiddenAtoms": [],
     "tokenUsage": {
-      "input_tokens": 5104,
+      "input_tokens": 5143,
       "cache_creation_input_tokens": 0,
       "cache_read_input_tokens": 13008,
       "cache_creation": {
         "ephemeral_5m_input_tokens": 0,
         "ephemeral_1h_input_tokens": 0
       },
-      "output_tokens": 373,
+      "output_tokens": 311,
       "service_tier": "standard",
       "inference_geo": "not_available"
     },
-    "costUSD": 0.0248094,
-    "elapsedSec": 9.5,
-    "generatedAt": "2026-07-10T04:22:53.357Z",
+    "costUSD": 0.023996399999999998,
+    "elapsedSec": 7.6,
+    "generatedAt": "2026-07-10T06:53:23.426Z",
     "model": "claude-sonnet-4-5-20250929"
   }
 }
