+++ sdf-js/examples/scaffold-pipeline/eval-funding-round/slots/slot-05-next-quarter.json
@@ -3,11 +3,11 @@
   "slotName": "next-quarter",
   "slotTitle": "Next Quarter",
   "slotPurpose": "Plan and priorities for the coming quarter",
-  "sourceSlideIdx": 13,
-  "sourceSlideTitle": "Strategic Outlook",
+  "sourceSlideIdx": 6,
+  "sourceSlideTitle": "AI-Driven Commerce Tools Development",
   "sceneData": {
     "name": "investor-update/next-quarter",
-    "layout": "grid",
+    "layout": "stage",
     "subjects": [
       {
         "type": "cover",
@@ -21,54 +21,45 @@
         }
       },
       {
-        "type": "kanban-board",
+        "type": "pillar-3up",
         "x": 40,
         "y": 160,
         "w": 1200,
         "h": 520,
         "args": {
-          "title": "Q4 Priorities",
-          "columns": [
+          "pillars": [
             {
-              "name": "Platform Evolution",
-              "cards": [
-                {
-                  "label": "Transition to comprehensive commerce platform",
-                  "sublabel": "Post-purchase specialist evolution"
-                },
-                {
-                  "label": "Multi-product adoption strategy",
-                  "sublabel": "Network effects activation"
-                }
+              "icon": "brain",
+              "heading": "AI-Driven Commerce Tools",
+              "body": "Development of next-generation shopping and marketing capabilities",
+              "accent": [
+                58,
+                138,
+                200
               ]
             },
             {
-              "name": "International Expansion",
-              "cards": [
-                {
-                  "label": "Global merchant network scaling",
-                  "sublabel": "ReturnBear acquisition integration"
-                },
-                {
-                  "label": "Beyond domestic market reach",
-                  "sublabel": "Returns infrastructure extension"
-                }
+              "icon": "briefcase",
+              "heading": "Capital Deployment",
+              "body": "New capital funds AI commerce tool development and market expansion",
+              "accent": [
+                100,
+                150,
+                200
               ]
             },
             {
-              "name": "AI Development",
-              "cards": [
-                {
-                  "label": "Three AI product initiatives",
-                  "sublabel": "Personalization and automation focus"
-                },
-                {
-                  "label": "AI-driven commerce tools",
-                  "sublabel": "Product roadmap execution"
-                }
+              "icon": "handshake",
+              "heading": "Merchant Engagement",
+              "body": "Suite designed to enhance merchant-customer engagement and conversion",
+              "accent": [
+                180,
+                210,
+                235
               ]
             }
-          ]
+          ],
+          "accentLine": true
         }
       }
     ]
@@ -84,20 +75,20 @@
     ],
     "forbiddenAtoms": [],
     "tokenUsage": {
-      "input_tokens": 5279,
+      "input_tokens": 5145,
       "cache_creation_input_tokens": 0,
       "cache_read_input_tokens": 13008,
       "cache_creation": {
         "ephemeral_5m_input_tokens": 0,
         "ephemeral_1h_input_tokens": 0
       },
-      "output_tokens": 449,
+      "output_tokens": 364,
       "service_tier": "standard",
       "inference_geo": "not_available"
     },
-    "costUSD": 0.026474400000000002,
-    "elapsedSec": 11,
-    "generatedAt": "2026-07-10T04:22:53.357Z",
+    "costUSD": 0.0247974,
+    "elapsedSec": 8.9,
+    "generatedAt": "2026-07-10T06:53:23.426Z",
     "model": "claude-sonnet-4-5-20250929"
   }
 }
